# DH.Curbside.API/App_Data/CurbsideApiSwaggerSpec.json
{  
   "swagger":"2.0",
   "info":{  
      "description":"This is a swagger spec API documentation for Curbside Application",
      "version":"1.0.0",
      "title":"Curbside API"
   },
...
   "paths":{  
...
      "/{tenantId}/whitelist":{  
         "get":{  
            "tags":[  
               "provisioning"
            ],
            "summary":"Get list of whitelisted users",
            "operationId":"getWhitelistUsers",
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"path",
                  "name":"tenantId",
                  "description":"organization guid",
                  "required":true,
                  "type":"string"
               },
               {  
                  "in":"query",
                  "name":"appId",
                  "description":"client app guid",
                  "required":true,
                  "type":"string",
                  "format":"uuid v4"
               }
            ],
            "responses":{  
               "200":{  
                  "description":"OK",
                  "schema":{  
                     "type":"array",
                     "items":{  
                        "$ref":"#definitions/User"
                     }
                  }
               }
            }
         }
      },
      "/{tenantId}/whitelist/{email}":{  
         "get":{  
            "tags":[  
               "provisioning"
            ],
            "summary":"Get user with whitelistinginfo",
            "operationId":"getWhitelistUser",
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"path",
                  "name":"tenantId",
                  "description":"organization guid",
                  "required":true,
                  "type":"string",
                  "format":"uuid v4"
               },
               {  
                  "in":"path",
                  "name":"email",
                  "description":"user email address",
                  "required":true,
                  "type":"string"
               }
            ],
            "responses":{  
               "200":{  
                  "description":"OK",
                  "schema":{  
                     "$ref":"#definitions/User"
                  }
               }
            }
         },
         "delete":{  
            "tags":[  
               "provisioning"
            ],
            "summary":"Deprovision an user from whitelist",
            "operationId":"deprovisionUser",
            "produces":[  
               "application/json"
            ],
            "parameters":[  
               {  
                  "in":"path",
                  "name":"tenantId",
                  "description":"organization guid",
                  "required":true,
                  "type":"string"
               },
               {  
                  "in":"path",
                  "name":"email",
                  "description":"user email address",
                  "required":true,
                  "type":"string"
               },
               {  
                  "in":"query",
                  "name":"appId",
                  "description":"submitter or reviewer app guid",
                  "required":true,
                  "type":"string",
                  "format":"uuid v4"
               }
            ],
            "responses":{  
               "200":{  
                  "description":"OK"
               },
               "404":{  
                  "description":"user or app doesn't exist"
               }
            }
         }
      },
...
   },
...
   "host":"virtserver.swaggerhub.com",
   "basePath":"/dignity/curbsideapi/1.0.0",
   "schemes":[  
      "https"
   ]
}